Refactor: Move component classes to separate files and update versioning
