+++ app/src/main/java/com/sq/thed_ck_licker/ecs/components/EffectStackComponent.kt
@@ -4,29 +4,12 @@ import android.util.Log
 import com.sq.thed_ck_licker.ecs.managers.ComponentManager
 import com.sq.thed_ck_licker.ecs.managers.EntityId
 import com.sq.thed_ck_licker.ecs.managers.hasComponent
-import javax.inject.Inject
 
 
-data class DrawDeckComponent @Inject constructor(
-    private val drawCardDeck: MutableList<Int>,
-) {
-    fun getSize(): Int {
-        return this.drawCardDeck.size
-    }
-
-    fun getDrawCardDeck(): MutableList<Int> {
-        return this.drawCardDeck
-    }
-
-    fun removeCards(cards: List<Int>) {
-        this.drawCardDeck.removeAll(cards)
-    }
-
-    fun removeCard(cardId: Int) {
-        this.drawCardDeck.remove(cardId)
-    }
-}
-
+/**
+ *  I am trying to decommission this one as of 0.1.2.133.
+ *  For substitute use onTurnStart + targetComponent.
+ */
 data class EffectStackComponent(private val effectEntities: MutableList<EntityId> = mutableListOf<EntityId>()) {
     fun getEffectEntities(): List<EntityId> {
         return this.effectEntities
@@ -44,7 +27,3 @@ data class EffectStackComponent(private val effectEntities: MutableList<EntityId
         this.effectEntities.add(effectEntity)
     }.invoke()
 }
-
-
-
-
